+++ Example/Sources/View Controllers/BasicExampleViewController.swift
@@ -33,6 +33,10 @@ class BasicExampleViewController: ChatViewController {
         messagesCollectionView.messagesLayoutDelegate = self
         messagesCollectionView.messagesDisplayDelegate = self
     }
+    
+    func textCellSizeCalculator(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> CellSizeCalculator? {
+        return nil
+    }
 }
 
 // MARK: - MessagesDisplayDelegate
@@ -136,5 +140,4 @@ extension BasicExampleViewController: MessagesLayoutDelegate {
     func messageBottomLabelHeight(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> CGFloat {
         return 16
     }
-    
 }